+++ README.md
@@ -126,11 +126,11 @@
 
 **POST** `/api/customer/carts`
 
-- [ ] 회원의 장바구니에 제품을 담는다.
+- [x] 회원의 장바구니에 제품을 담는다.
 
 `예외`
 
-- [ ] id 에 해당하는 제품이 없는 경우 - **404**
+- [x] id 에 해당하는 제품이 없는 경우 - **404**
 
 <br>
