+++ src/main/java/woowacourse/shoppingcart/exception/notfound/InvalidProductException.java
@@ -1,6 +1,6 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.notfound;
 
-public class InvalidProductException extends DomainException {
+public class InvalidProductException extends NotFoundException {
 
     public InvalidProductException() {
         this("올바르지 않은 사용자 아이디이거나 상품 아이디 입니다.");
