+++ src/test/java/woowacourse/shoppingcart/ui/CartItemControllerTest.java
@@ -1,6 +1,7 @@
 package woowacourse.shoppingcart.ui;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -16,6 +17,7 @@ import woowacourse.shoppingcart.application.dto.CustomerSaveServiceRequest;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.dto.AddCartItemRequest;
+import woowacourse.shoppingcart.exception.notfound.NotFoundException;
 
 @SpringBootTest
 @Sql("classpath:resetTables.sql")
@@ -42,6 +44,17 @@ class CartItemControllerTest {
 
         // then
         assertThat(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 제품 id로 항목 추가하면 예외 발생하는 기능")
+    void addCartItem_invalidProductId_throwsException() {
+        // given
+        customerService.save(new CustomerSaveServiceRequest("klay", "email@gmail.com", "12345678"));
+
+        // when, then
+        assertThatThrownBy(() -> cartItemController.addCartItem(1L, new AddCartItemRequest(1L)))
+                .isInstanceOf(NotFoundException.class);
     }
 
     @Test
